--- a/src/blockchain_utilities/blockchain_ancestry.cpp
+++ b/src/blockchain_utilities/blockchain_ancestry.cpp
@@ -34,7 +34,7 @@
 #include <boost/archive/portable_binary_oarchive.hpp>
 #include <boost/filesystem/path.hpp>
 #include "common/unordered_containers_boost_serialization.h"
-#include "common/command_line.h"
+#include "common/command_line_functions.h"
 #include "common/varint.h"
 #include "cryptonote_basic/cryptonote_boost_serialization.h"
 #include "cryptonote_core/tx_pool.h"