--- a/src/blockchain_utilities/blockchain_export.cpp
+++ b/src/blockchain_utilities/blockchain_export.cpp
@@ -28,7 +28,7 @@
 
 #include "bootstrap_file.h"
 #include "blocksdat_file.h"
-#include "common/command_line.h"
+#include "common/command_line_functions.h"
 #include "cryptonote_core/tx_pool.h"
 #include "cryptonote_core/cryptonote_core.h"
 #include "blockchain_db/blockchain_db.h"