--- a/src/common/command_line.h
+++ b/src/common/command_line.h
@@ -25,21 +25,15 @@
 // INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
 // STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF
 // THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-// 
-// Parts of this file are originally copyright (c) 2012-2013 The Cryptonote developers
 
 #pragma once
 
-#include <functional>
-#include <iostream>
-#include <sstream>
-#include <array>
-#include <type_traits>
-
-#include <boost/program_options/parsers.hpp>
-#include <boost/program_options/options_description.hpp>
-#include <boost/program_options/variables_map.hpp>
 #include "include_base_utils.h"
+
+#include "fwd/boost_monero_program_options_fwd.h"
+
+#include <functional>
+#include <array>
 
 namespace command_line
 {
@@ -114,185 +108,6 @@ namespace command_line
 
     bool not_use_default;
   };
-
-  template<typename T>
-  boost::program_options::typed_value<T, char>* make_semantic(const arg_descriptor<T, true>& /*arg*/)
-  {
-    return boost::program_options::value<T>()->required();
-  }
-
-  template<typename T>
-  boost::program_options::typed_value<T, char>* make_semantic(const arg_descriptor<T, false>& arg)
-  {
-    auto semantic = boost::program_options::value<T>();
-    if (!arg.not_use_default)
-      semantic->default_value(arg.default_value);
-    return semantic;
-  }
-
-  template<typename T>
-  boost::program_options::typed_value<T, char>* make_semantic(const arg_descriptor<T, false, true>& arg)
-  {
-    auto semantic = boost::program_options::value<T>();
-    if (!arg.not_use_default) {
-      std::ostringstream format;
-      format << arg.depf(false, true, arg.default_value) << ", "
-             << arg.depf(true, true, arg.default_value) << " if '"
-             << arg.ref.name << "'";
-      semantic->default_value(arg.depf(arg.ref.default_value, true, arg.default_value), format.str());
-    }
-    return semantic;
-  }
-
-  template<typename T, int NUM_DEPS>
-  boost::program_options::typed_value<T, char>* make_semantic(const arg_descriptor<T, false, true, NUM_DEPS>& arg)
-  {
-    auto semantic = boost::program_options::value<T>();
-    if (!arg.not_use_default) {
-      std::array<bool, NUM_DEPS> depval;
-      depval.fill(false);
-      std::ostringstream format;
-      format << arg.depf(depval, true, arg.default_value);
-      for (size_t i = 0; i < depval.size(); ++i)
-      {
-        depval.fill(false);
-        depval[i] = true;
-        format << ", " << arg.depf(depval, true, arg.default_value) << " if '" << arg.ref[i]->name << "'";
-      }
-      for (size_t i = 0; i < depval.size(); ++i)
-        depval[i] = arg.ref[i]->default_value;
-      semantic->default_value(arg.depf(depval, true, arg.default_value), format.str());
-    }
-    return semantic;
-  }
-
-  template<typename T>
-  boost::program_options::typed_value<T, char>* make_semantic(const arg_descriptor<T, false>& arg, const T& def)
-  {
-    auto semantic = boost::program_options::value<T>();
-    if (!arg.not_use_default)
-      semantic->default_value(def);
-    return semantic;
-  }
-
-  template<typename T>
-  boost::program_options::typed_value<std::vector<T>, char>* make_semantic(const arg_descriptor<std::vector<T>, false>& /*arg*/)
-  {
-    auto semantic = boost::program_options::value< std::vector<T> >();
-    semantic->default_value(std::vector<T>(), "");
-    return semantic;
-  }
-
-  template<typename T, bool required, bool dependent, int NUM_DEPS>
-  void add_arg(boost::program_options::options_description& description, const arg_descriptor<T, required, dependent, NUM_DEPS>& arg, bool unique = true)
-  {
-    if (0 != description.find_nothrow(arg.name, false))
-    {
-      CHECK_AND_ASSERT_MES(!unique, void(), "Argument already exists: " << arg.name);
-      return;
-    }
-
-    description.add_options()(arg.name, make_semantic(arg), arg.description);
-  }
-
-  template<typename T>
-  void add_arg(boost::program_options::options_description& description, const arg_descriptor<T, false>& arg, const T& def, bool unique = true)
-  {
-    if (0 != description.find_nothrow(arg.name, false))
-    {
-      CHECK_AND_ASSERT_MES(!unique, void(), "Argument already exists: " << arg.name);
-      return;
-    }
-
-    description.add_options()(arg.name, make_semantic(arg, def), arg.description);
-  }
-
-  template<>
-  inline void add_arg(boost::program_options::options_description& description, const arg_descriptor<bool, false>& arg, bool unique)
-  {
-    if (0 != description.find_nothrow(arg.name, false))
-    {
-      CHECK_AND_ASSERT_MES(!unique, void(), "Argument already exists: " << arg.name);
-      return;
-    }
-
-    description.add_options()(arg.name, boost::program_options::bool_switch(), arg.description);
-  }
-
-  template<typename charT>
-  boost::program_options::basic_parsed_options<charT> parse_command_line(int argc, const charT* const argv[],
-    const boost::program_options::options_description& desc, bool allow_unregistered = false)
-  {
-    auto parser = boost::program_options::command_line_parser(argc, argv);
-    parser.options(desc);
-    if (allow_unregistered)
-    {
-      parser.allow_unregistered();
-    }
-    return parser.run();
-  }
-
-  template<typename F>
-  bool handle_error_helper(const boost::program_options::options_description& desc, F parser)
-  {
-    try
-    {
-      return parser();
-    }
-    catch (const std::exception& e)
-    {
-      std::cerr << "Failed to parse arguments: " << e.what() << std::endl;
-      std::cerr << desc << std::endl;
-      return false;
-    }
-    catch (...)
-    {
-      std::cerr << "Failed to parse arguments: unknown exception" << std::endl;
-      std::cerr << desc << std::endl;
-      return false;
-    }
-  }
-
-  template<typename T, bool required, bool dependent, int NUM_DEPS>
-  typename std::enable_if<!std::is_same<T, bool>::value, bool>::type has_arg(const boost::program_options::variables_map& vm, const arg_descriptor<T, required, dependent, NUM_DEPS>& arg)
-  {
-    auto value = vm[arg.name];
-    return !value.empty();
-  }
-
-  template<typename T, bool required, bool dependent, int NUM_DEPS>
-  bool is_arg_defaulted(const boost::program_options::variables_map& vm, const arg_descriptor<T, required, dependent, NUM_DEPS>& arg)
-  {
-    return vm[arg.name].defaulted();
-  }
-
-  template<typename T>
-  T get_arg(const boost::program_options::variables_map& vm, const arg_descriptor<T, false, true>& arg)
-  {
-    return arg.depf(get_arg(vm, arg.ref), is_arg_defaulted(vm, arg), vm[arg.name].template as<T>());
-  }
-
-  template<typename T, int NUM_DEPS>
-  T get_arg(const boost::program_options::variables_map& vm, const arg_descriptor<T, false, true, NUM_DEPS>& arg)
-  {
-    std::array<bool, NUM_DEPS> depval;
-    for (size_t i = 0; i < depval.size(); ++i)
-      depval[i] = get_arg(vm, *arg.ref[i]);
-    return arg.depf(depval, is_arg_defaulted(vm, arg), vm[arg.name].template as<T>());
-  }
-
-  template<typename T, bool required>
-  T get_arg(const boost::program_options::variables_map& vm, const arg_descriptor<T, required>& arg)
-  {
-    return vm[arg.name].template as<T>();
-  }
- 
-  template<bool dependent, int NUM_DEPS>
-  inline bool has_arg(const boost::program_options::variables_map& vm, const arg_descriptor<bool, false, dependent, NUM_DEPS>& arg)
-  {
-    return get_arg(vm, arg);
-  }
-
 
   extern const arg_descriptor<bool> arg_help;
   extern const arg_descriptor<bool> arg_version;