Got rid of program_options from headers and limited them in common/command_line_functions.h and tests/core_tests/chaingen_program_options.h, which are included only by .cpp files, thus preventing from spamming the project with program_options, where they are not needed.

The change is transactional: many files were touched, but cutting slow (boost) headers from top level project headers requires that ALL of the occurrences from the sub-headers are moved. If just one is left out, then the entire effort is meaningless, as the slow header will be included by the users of the top level headers anyway.
To satisfy the above requirement and keep the PR as small as possible in the same time, the strategy that I took, is cutting one boost header library at a time (ie. the whole program_options library).

I have introduced a new target: "fwd", where I will store all kind of lightweight forward declarations in a hierarchical manner.
Below are the files including the fwd header now. Directly or indirectly:
![boost__monero__program__options__fwd_8h__dep__incl](https://user-images.githubusercontent.com/63722585/97739295-f7a8dc00-1adf-11eb-808d-3179a1059ab2.png)
Some still have to include the originals, but this is reasonably limited:
![image](https://user-images.githubusercontent.com/63722585/97774482-6c1f6180-1b58-11eb-8b1d-2a0ca04f54b8.png)
Compare with the previous state:
![image](https://user-images.githubusercontent.com/63722585/97774505-b6a0de00-1b58-11eb-82f6-dd2606e57a8b.png)


**Verification of the results**
Touching the files including the thinned down includes and rebuilding with ccache disabled delivered the following results:
before: 37m33,285s
after:    36m17,847s
This is a speed up of 3.3% for the relevant files, which now didn't have to compile the boost/program_options anymore.
Generating the list of files can be done on my branch with: 
egrep -lir  "common/command_line.h" .
To verify whether there's any boost_options header left included in the headers, please execute on my branch (and compare on the original):
egrep -ir --include=*.{h,hpp,cpp,inl} "boost/program_options" .
This is true for every header, except command_line_functions.h, a new header with templates, where it wasn't possible. The scope of command_line_functions.h was limited to as much as it was possible.

Summary:
| Master   |      Patch      | Change |
|----------|-------------|-------------|
| Parsing (frontend):         1611.3 s |     Parsing (frontend):         1565.1 s   |  -2.87% | 
| Codegen & opts (backend):   1663.8 s | Codegen & opts (backend): 1626.5  | -2.24% |

Active work time (including tools preparation):
4.2h
Passive work time (test and verification of results):
0.6h test
0.9h verification
2.1h additional 2 verifications for the review