@mj-xmr 2020-10-24
> That autoformatter stuff is just injecting noise into this large set of patches. Please omit that.

OK, I'm undoing it.