@mj-xmr 2020-10-27
> There's a number of header files that reference `boost::program_options`, but the header has no direct declaration or definition and is relying on indirect includes. This is somewhat matter of taste, but now manipulating other headers breaks these headers. Typically this is handled by just forward declaring directly in the header making the reference ... I stopped commenting on all of these cases. The ones involved `common/command_line.h` are probably less squeamish, but there's a number of indirect includes through `cryptonote_basic.h` or something.

OK, point taken. It's not a matter of taste if it involves more work from others and I want this PR to do it once and for all.
Creating a proper fwd header and using (including) it whenever the opts are referenced sound like a plan.