@mj-xmr 2020-10-30
I'm putting this on hold for now, as I apparently need to introduce specialized tooling, in order to do things right and in a more targeted manner.
I'm strongly considering:
https://github.com/Leedehai/C-include-2-dot