@mj-xmr 2021-05-27
Related to https://github.com/monero-project/monero/pull/7735
I need to setup a proper framework for testing compatibility between various boost versions. 

Putting to a draft more until then.